Implement text mode. It's limited to 8x8 since that's the only font size VICKYII supports. Had to rework a bit how the vt52/conout work. More work would be needed there to be able to use various modes (frame buffer, text mode) as currently it's all driven by #if.
The computation of the number of columns according to VICKY border size is dodgy. I changed the blinking mode and I have no idea if it still work for framebuffer modes, that needs testing.

Character 0 is used for the cursor.
Flashing and blinking rates are not really controllable because VICKY only has 4 flash speeds.